# spec/swagger.yaml
definitions:
...
  handlers.ContainerGPURequest:
    properties:
      devices:
        items:
          type: string
        type: array
    type: object
...
  handlers.CreateContainerRequest:
    properties:
      gpu:
        $ref: '#/definitions/handlers.ContainerGPURequest'
      image:
        type: string
      local_workspace_path:
        type: string
      restore_data:
        type: boolean
      snapshotter:
        type: string
      workspace_backend:
        type: string
    type: object
  handlers.CreateContainerResponse:
    properties:
      cdi_devices:
        items:
          type: string
        type: array
      container_id:
        type: string
      container_path:
        type: string
      data_restored:
        type: boolean
      has_preserved_data:
        type: boolean
      image:
        type: string
      snapshotter:
        type: string
      started:
        type: boolean
      workspace_backend:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.GetContainerResponse:
    properties:
      cdi_devices:
        items:
          type: string
        type: array
      container_id:
        type: string
      container_path:
        type: string
      created_at:
        type: string
      has_preserved_data:
        type: boolean
      image:
        type: string
      legacy:
        type: boolean
      namespace:
        type: string
      status:
        type: string
      task_running:
        type: boolean
      updated_at:
        type: string
      workspace_backend:
        type: string
    type: object
...
  handlers.terminalInfoResponse:
    properties:
      available:
        type: boolean
      shell:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/container:
    delete:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Export /data before deletion
        in: query
        name: preserve_data
        type: boolean
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete MCP container for bot
      tags:
      - workspace
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetContainerResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get container info for bot
      tags:
      - workspace
    post:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Create container payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateContainerRequest'
      responses:
        "200":
          description: SSE stream of container creation events
          schema:
            $ref: '#/definitions/handlers.CreateContainerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create and start MCP container for bot
      tags:
      - workspace
...
  /bots/{bot_id}/container/terminal:
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.terminalInfoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Check terminal availability for bot container
      tags:
      - workspace
  /bots/{bot_id}/container/terminal/ws:
    get:
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - default: 80
        description: Initial terminal columns
        in: query
        name: cols
        type: integer
      - default: 24
        description: Initial terminal rows
        in: query
        name: rows
        type: integer
      - description: Auth token
        in: query
        name: token
        type: string
      responses:
        "101":
          description: WebSocket upgrade
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Interactive WebSocket terminal for bot container
      tags:
      - workspace
...
swagger: "2.0"